refactor: move createStatistic fn to new file to fix circular dependency issue
